Renamed LinkedList class file and restructured the class code
The Node class is now a part of the LinkedList class. Cleaned up some comments
